+++ apps/components/Img_Factory/imgfactory.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-#this belongs in apps/components/Img_Factory/imgfactory.py - Version: 100
+#this belongs in apps/components/Img_Factory/imgfactory.py - Version: 101
 # X-Seti - Feb 24 2026 - IMG Factory 1.6 - Icon system, button layout
 
 """